% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antagonism_aggregate.R
\name{aggregate_and_prioritize}
\alias{aggregate_and_prioritize}
\title{Aggregate and prioritize compounds}
\usage{
aggregate_and_prioritize(
  dfs.to.process = "results/GTP_CDR/.+_.+_all\\\\.signatures\\\\.AvgRank\\\\.csv\\\\.gz$",
  limit.dfs.to = "trt_cp|trt_sh|trt_oe|trt_xpr",
  limit.models.to = NA,
  limit.gwass.to = NA,
  output.dir = "results/GTP_CDR/",
  ref.drug = TRT_CP.INFO.20200324,
  ref.cell = CELL.LINE.INFO,
  iterative.tissue = TRUE,
  iterative.aggregate = TRUE,
  min.experiments.n = 2,
  n.cores = parallel::detectCores() - 2
)
}
\arguments{
\item{dfs.to.process}{dfs to be loaded based on regular expression.}

\item{limit.dfs.to}{regular expression for inclusion rlist of datasets.}

\item{limit.models.to}{Limit model_IDs}

\item{limit.gwass.to}{Limit gwas}

\item{output.dir}{Default is "results/CDR_GTP/"}

\item{ref.drug}{File with information for compounds (trt_cp) such as mechanism of action, indication etc. Look at data-raw if you want to generate a similar file}

\item{ref.cell}{File with informaiton about the cell lines used in the perturbagen library.}

\item{iterative.tissue}{when having more than one tissue/cell type run for each tissue or cell types. Default is FALSE}

\item{iterative.aggregate}{when running iterative also run the aggregate of all tissues or cell types. Default is TRUE}

\item{min.experiments.n}{Min experiment per perturbagen (default is 2)}

\item{n.cores}{Number of cores to use if multithreaded}

\item{compound.type}{Options are CDR for computational drug repurposing/repositioning and GTP for gene target prioritization which considers trt_sh (shRNA), trt_oe ().}

\item{drug.file}{trt_cp specirific parametereCompound file; default is repurposing_drugs_20200324.txt from clue.io}

\item{cell.file}{Compound file; default is cell_lkines.csv from clue.io}
}
\value{
Compound level summary
}
\description{
Aggregates across models. Run separately if you wish to separate models. Perturbagen Library used. We use the LINKS Phase II L1000 dataset (GSE70138) perturbagen reference library1. All inferred genes (AIG; n=12,328) are considered. Only “gold” signatures are considered. Imputed transcriptomes used. Analysis is limited to 17 imputed transcriptomes: (1) the B2 phenotype, and (2) EpiXcan tissue models that have at least one FDR-significant finding (FDR significance takes into account all COVID phenotypes and all 42 tissue models considered): Adipose: subcutaneous (GTEx), Adipose: subcutaneous (STARNET), Adipose: visceral (GTEx), Adipose: visceral (STARNET), Artery: Aorta (GTEx), Artery: Aorta (STARNET), Artery: Mammary (STARNET), Blood (STARNET), GI: esophagus, GE junction (GTEx), GI: esophagus, mucosa (GTEx), GI: muscularis (GTEx), GI: pancreas (GTEx), Muscle: skeletal (GTEx), Muscle: skeletal (STARNET), Reproductive: mammary tissue (GTEx), Respiratory: lung (GTEx), Skin: sun exposed lower leg (GTEx). Summarization and prioritization approach: From the original CDR pipeline2 (using 100 permutations), applied to the imputed transcriptome of each tissue, we obtain the average rank of the compound in antagonizing the GReX and the permutation p values. After pulling together all the results, we perform a Mann-Whithney U test for each candidate compound/ shRNA against all other compounds to see if the candidate’s rankings significantly deviate from the median rank. For each candidate we also estimate a GReX antagonism pseudo z-score, which is defined as the negative Hodges-Lehmann estimator (of the median difference between that specific candidate vs. the other candidates) divided by the standard deviation of the ranks of the compounds (-Hodges-Lehmann estimatorperturbagenSD average ranks of all perturbagens); a positive pseudo z-score is interpreted as a potential therapeutic candidate whereas a negative pseudo z-score would suggest that the shRNA is not antagonizing the imputed transcriptome and is thus likely to exacerbate the phenotype. Of note is that at this stage each candidate is compared against the other candidates but we can confirm that the candidate is effectively antagonizing the GReX by looking at the original permutation p values. FDR is estimated with the Benjamini–Hochberg procedure3. Additional information for chemical compound analyses: Analysis is limited to compounds eligible for drug repurposing (n=495). Drug information for the compounds under consideration (e.g. clinical phase, mechanism of action and molecular targets) was obtained from http://www.broadinstitute.org/repurposing (file date: 3/24/2020). For comparison with other studies; the compounds under question were compared with all the other compounds. For the mechanism of action comparison all compounds with a known mechanism of action represented with two or more candidates are considered. Final recommendations are for launched medications and FDR correction is applied only to launched compounds. Additional information for shRNA analyses: All shRNAs were considered.
To run models separately a loop will have to pass arguments to limit.models.to (and keep NA if needed). Requires up to 10GB per thread.
Consider "clinical_phase" column options are "Launched", "Phase 3", "Phase 2/Phase 3", "Phase 2", "Phase 1/Phase 2", "Phase 1", "Preclinical" and "Withdrawn". A sensible list for CDR is "Launched|Phase 3|Phase 2"
}
