% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_five_rank_method.R
\name{five_rank_method}
\alias{five_rank_method}
\title{Distilled main function for five-rank method approach by So et al 2017}
\usage{
five_rank_method(thres.N, x, scramble = F)
}
\arguments{
\item{thres.N}{threshold NA is all values otherwise consider top values}

\item{x}{payload of interest}

\item{scramble}{scramble the z-scores (used for permutation analusis)}
}
\value{
Returns five ranks
}
\description{
Intention is to be used to run the 5 methods for each GWAS - model_ID combination.
The following 5 variables/methods need to be saved in splitting the drug matrices:
1. cor.pearson : Pearson (all)
2. cor.spearman : Spearman (all)
3. extreme.cor.spearman : Spearman (most differentially expressed genes)
4. extreme.cor.pearson : Pearson (most differentially expressed genes)
5. ks.signed : KS method (most differentially expressed genes)
This is a multithreaded data.table based re-write of So HC et al. Analysis of
 genome-wide association data highlights candidates for drug repositioning in
 psychiatry. Nat Neurosci. 2017 Oct;20(10):1342-1349. PMID: 28805813.
It has been adapted to run for signatures (individual parameters) vs.
summarized compound scores.
}
\keyword{five}
\keyword{method}
\keyword{rank}
