% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figs_gtp_prioritization.R
\name{gtp_pvalue_qqplot}
\alias{gtp_pvalue_qqplot}
\title{pvalue_qqplot}
\usage{
gtp_pvalue_qqplot(
  ps = ppoints(1:10),
  ci = 0.95,
  type = c("pvalue", "zscore"),
  stat.prefix = "",
  ppoints.a = NA,
  figsDir = "figs/",
  figure.name = "pvalue_qqplot",
  save_plot = F
)
}
\arguments{
\item{ps}{Vector of p-values or z-scores.}

\item{ci}{Size of the confidence interval, 95% by default.}

\item{type}{"pvalue" or "zscore" accepted for the provided statistic}

\item{stat.prefix}{text for the method or title of the qqplot}

\item{figsDir}{Default is "figs/"}

\item{figure.name}{Default is "qqplot_TWAS_vs_FunWAS"}

\item{save_plot}{Save plots in various useful formats}
}
\value{
returns ggplot graph
}
\description{
Create a quantile-quantile plot with ggplot2. Script was adapted from: https://slowkow.com/notes/ggplot2-qqplot/. The null is the ordinates for probability plotting Generates the sequence of probability points (1:m - a)/(m + (1-a)-a) as generated by ppoints similarly to the one used in qqplot and qqnorm.
}
\details{
Future improvements
- Stouffer's method or other zscore based meta-analysis
- Scaling prior to getting z combined


Assumptions:
  - Expected P values are uniformly distributed.
  - Confidence intervals assume independence between tests.
    We expect deviations past the confidence intervals if the tests are
    not independent.
    For example, in a genome-wide association study, the genotype at any
    position is correlated to nearby positions. Tests of nearby genotypes
    will result in similar test statistics.
}
\examples{
library(ggplot2)
gg_qqplot(runif(1e2)) + theme_grey(base_size = 24)
}
