% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setttings_parse_recipe.R
\name{parse_recipe}
\alias{parse_recipe}
\title{Parse LSF project recipe file}
\usage{
parse_recipe(recipe.file)
}
\arguments{
\item{recipe.file}{recipe file}
}
\value{
A list file with all included parameters
}
\description{
Parse LSF project recipe file
}
\keyword{LSF}
\keyword{antagonism}
\keyword{step1}
