% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antagonism_perform.R
\name{perform_antagonism}
\alias{perform_antagonism}
\title{Antagonism core script}
\usage{
perform_antagonism(
  df = paste0("output/2.METAXCAN/", basename(getwd()),
    "_df.all.annotated.onlytranscripts.csv.gz"),
  column.feature = "feature",
  column.statistic = "zscore",
  column.trait = "gwas",
  column.source = "model_ID",
  results.dir = "results/GTP_CDR/",
  n.threads = parallel::detectCores() - 2,
  signature.dir =
    "/sc/arion/projects/va-biobank/resources/CMap/cmap_l1000_2021-11-20/eachDrug/",
  gene.anno.file =
    "/sc/arion/projects/va-biobank/resources/CMap/cmap_l1000_2021-11-20/geneinfo_beta.txt",
  grep.sig.pattern = "trt_cp|trt_sh|trt_oe|trt_xpr",
  noperm = 100,
  thres.N.vector = c(50, 100, 250, 500),
  sig.annotation = SIG.INFO.20211120,
  overwrite.intermediate = TRUE,
  model.banlist.grep = ":: Transcripts ::|:: H3K4me3 ::|:: H3K27ac ::|:: CA ::",
  prototyping = NA
)
}
\arguments{
\item{df}{twas data frame}

\item{column.feature}{expected input is ENSEMBL ID and currently only works with genes (default: feature)}

\item{column.statistic}{z-score statistic (default: zscore)}

\item{column.trait}{trait name (default: gwas)}

\item{column.source}{disease signature source, e.g. microglia, DLPFC, meta-analysis, etc. (default: model_ID). For non-twas inputs make sure there is a column describing the input.}

\item{results.dir}{Results main dir. Subdirectories will be created}

\item{n.threads}{number f threads for multicore strategy, default is detectCores()-2}

\item{signature.dir}{Location of signature files. Consider saving locally to speed things up (e.g. /scratch/cmap_l1000_2021_01_28/). Look into the split_gctx on how to generate these files.}

\item{gene.anno.file}{Gene annotation file (provided from L1000)}

\item{grep.sig.pattern}{Regular exrpression pattern to grep from signature name files. Default is to process sets for computational drug repurposing and gene target prioritization.}

\item{noperm}{Number of permutations (the final number of permutation data points is noperm × number of drugs)}

\item{thres.N.vector}{defining the threshold(s) K such that only top K items are included for comparison of expressions: use the same thresholds as So et al.}

\item{sig.annotation}{signature annotation. Default is SIG.INFO.20211120}

\item{overwrite.intermediate}{default is not to overwrite the all.signatures.csv.gz (FALSE). This doesn't change from analysis to analysis.}

\item{model.banlist.grep}{regular expression to exclude non-gene-level models. Default is: ":: Transcripts ::|:: H3K4me3 ::|:: H3K27ac ::|:: CA ::"}

\item{prototyping}{number of signature files (each one has around 300) to test again for pipeline testing (default is NA which is all)}
}
\value{
N/A. Saves the figures
}
\description{
This is intended to run in your personal computer. There is a version that is optimized for LSF based clusters.
Runs the 5 methods for each GWAS / model_ID combination.
The following 5 variables/methods need to be saved in splitting the drug matrices:
1. cor.pearson : Pearson (all)
2. cor.spearman : Spearman (all)
3. extreme.cor.spearman : Spearman (most differentially expressed genes)
4. extreme.cor.pearson : Pearson (most differentially expressed genes)
5. ks.signed : KS method (most differentially expressed genes)
This is a multithreaded data.table based re-write of So HC et al. Analysis of
 genome-wide association data highlights candidates for drug repositioning in
 psychiatry. Nat Neurosci. 2017 Oct;20(10):1342-1349. PMID: 28805813.
Analysis is run for each GWAS (trait) - model_ID (imputation model or specific analysis) pair
It has been adapted to run for signatures (individual parameters) vs.
summarized compound scores.
}
\keyword{miami}
\keyword{plot}
\keyword{regional}
\keyword{visualization}
