% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antagonism_perform_lsf_S01_wrapper.R
\name{perform_antagonism_lsf_S01_wrapper}
\alias{perform_antagonism_lsf_S01_wrapper}
\title{Antagonism core script}
\usage{
perform_antagonism_lsf_S01_wrapper(
  working.directory,
  recipe.file,
  dryrun = FALSE,
  df,
  column.feature = "feature",
  column.statistic = "zscore",
  column.trait = "gwas",
  column.source = "model_ID",
  results.dir = "results/GTP_CDR/",
  n.threads = parallel::detectCores() - 2,
  signature.dir =
    "/sc/arion/projects/va-biobank/resources/CMap/cmap_l1000_2021-11-20/eachDrug/",
  gene.anno.file =
    "/sc/arion/projects/va-biobank/resources/CMap/cmap_l1000_2021-11-20/geneinfo_beta.txt",
  grep.sig.pattern = "trt_cp|trt_sh|trt_oe|trt_xpr",
  noperm = 100,
  thres.N.vector = c(50, 100, 250, 500),
  sig.annotation = SIG.INFO.20211120,
  overwrite.intermediate = TRUE,
  model.banlist.grep = ":: Transcripts ::|:: H3K4me3 ::|:: H3K27ac ::|:: CA ::",
  prototyping = NA
)
}
\arguments{
\item{working.directory}{provide working directory}

\item{recipe.file}{provide recipe.file location}

\item{dryrun}{if true the jobs are not submitted}

\item{df}{twas data frame}

\item{column.feature}{expected input is ENSEMBL ID and currently only works with genes (default: feature)}

\item{column.statistic}{z-score statistic (default: zscore)}

\item{column.trait}{trait name (default: gwas)}

\item{column.source}{disease signature source, e.g. microglia, DLPFC, meta-analysis, etc. (default: model_ID)}

\item{results.dir}{Results main dir. Subdirectories will be created}

\item{n.threads}{number f threads for multicore strategy, default is detectCores()-2}

\item{signature.dir}{Location of signature files. Consider saving locally to speed things up (e.g. /scratch/cmap_l1000_2021_01_28/). Look into the split_gctx on how to generate these files.}

\item{gene.anno.file}{Gene annotation file (provided from L1000)}

\item{grep.sig.pattern}{Regular exrpression pattern to grep from signature name files. Default is to process sets for computational drug repurposing and gene target prioritization.}

\item{noperm}{Number of permutations (the final number of permutation data points is noperm × number of drugs)}

\item{thres.N.vector}{defining the threshold(s) K such that only top K items are included for comparison of expressions: use the same thresholds as So et al.}

\item{sig.annotation}{signature annotation. Default is SIG.INFO.20211120}

\item{overwrite.intermediate}{default is not to overwrite the all.signatures.csv.gz (FALSE). This doesn't change from analysis to analysis.}

\item{model.banlist.grep}{regular expression to exclude non-gene-level models. Default is: ":: Transcripts ::|:: H3K4me3 ::|:: H3K27ac ::|:: CA ::"}

\item{prototyping}{number of signature files (each one has around 300) to test again for pipeline testing (default is NA which is all)}
}
\value{
N/A. Saves required files
}
\description{
Called by:
- exec/antagonist_S01_wrapper.R
Calls:
- exec/antagonist_S01PA_core.R
This is the first part of step 1 of the pipeline. The file with the disease
signature is provided and may have multiple traits (e.g. TWASs) and sources
(e.g. tissues). The parallelization at this step is at the level of the
signature.
}
\details{
General information below:
This is intended to run in an LSF custer. There is a version that is optimized for running locally.
Runs the 5 methods for each GWAS / model_ID combination.
The following 5 variables/methods need to be saved in splitting the drug matrices:
1. cor.pearson : Pearson (all)
2. cor.spearman : Spearman (all)
3. extreme.cor.spearman : Spearman (most differentially expressed genes)
4. extreme.cor.pearson : Pearson (most differentially expressed genes)
5. ks.signed : KS method (most differentially expressed genes)
This is a multithreaded data.table based re-write of So HC et al. Analysis of
 genome-wide association data highlights candidates for drug repositioning in
 psychiatry. Nat Neurosci. 2017 Oct;20(10):1342-1349. PMID: 28805813.
It has been adapted to run for signatures (individual parameters) vs.
summarized compound scores.
}
\keyword{LSF}
\keyword{antagonism}
\keyword{step1}
