% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_random_gene_name.R
\name{random_gene_name}
\alias{random_gene_name}
\title{Generate random gene names}
\usage{
random_gene_name(how.many.samples = 100, how.many.letters = 3, thisseed = NULL)
}
\arguments{
\item{how.many.samples}{how many random genes to return}

\item{how.many.letters}{how many letters each gene}

\item{thisseed}{can use seed to avoid occurance of inappropriate words}
}
\value{
returns vector
}
\description{
Generate random gene names
}
