% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showcase.R
\name{showcase_method_cdr_gtp}
\alias{showcase_method_cdr_gtp}
\title{Showcase how the method works}
\usage{
showcase_method_cdr_gtp(
  perturbagen,
  twas.model = NA,
  twas.trait = NA,
  perturbagen.sig_id = NA,
  twas = paste0("output/2.METAXCAN/", basename(getwd()),
    "_df.all.annotated.onlytranscripts.csv.gz"),
  twas.feature.column.name = "feature",
  twas.feature.name.column.name = "feature_name",
  twas.zscore.column.name = "zscore",
  twas.model.column.name = "model_ID",
  twas.trait.column.name = "gwas",
  sig.info = SIG.INFO.20211120,
  gtp.cdr.dir = "results/GTP_CDR/",
  gene.anno.file =
    "/sc/arion/projects/va-biobank/resources/CMap/cmap_l1000_2021-11-20/geneinfo_beta.txt"
)
}
\arguments{
\item{perturbagen}{pert_iname; e.g. "benactyzine"}

\item{twas.model}{GFI model name  (can also be DGE experiment type or cell type)}

\item{twas.trait}{Trait or GWAS}

\item{twas}{Disease signature data frame or file (default is the output of the MultiWAS pipeline)}

\item{twas.feature.column.name}{genomic feature column name (default: feature). Expects ENSEMBL IDs and filters out everything that doesn't start with ENSG.}

\item{twas.feature.name.column.name}{genomic feature column name (default: feature_name). Friendly name}

\item{twas.zscore.column.name}{z-score statistic column name (default: zscore)}

\item{twas.model.column.name}{GFI model (can also be DGE experiment type or cell type) column name (default: model_ID)}

\item{twas.trait.column.name}{Trait or GWAS column name (default: gwas)}

\item{sig.info}{File with perturbagen signature information (default: SIG.INFO.20211120)}

\item{gtp.cdr.dir}{Directory where GTP/CDR saves its results (default: "results/GTP_CDR/")}

\item{gene.anno.file}{Gene annotation file (provided from L1000)}
}
\value{
N/A. Saves the figures and csv file
}
\description{
This script generates most of the figures needed for the Fig. S5 in [1]. As a
reminder the core method is the 5-rank method[2]. This function requires that
GTP or CDR has run already.
- Panel A: Scatter plot of disease and perturbagen effects (signature level)
- Panel B: Density plot of Signatures for perturbagen vs. baseline.
- Panel C:
   - Left: Scatter plot of disease and perturbagen effects (compound level)
   - Middle: histogram of combined z for all shRNAs (compound level)
   - Right: Combined z -score for prioritization.
}
\details{
1. Voloudakis, G., Vicari, J.M., Venkatesh, S. et al. A translational genomics
approach identifies IL10RB as the top candidate gene target for COVID-19
susceptibility. npj Genom. Med. 7, 52 (2022).
https://doi.org/10.1038/s41525-022-00324-x

2. So HC et al. Analysis of genome-wide association data highlights
candidates for drug repositioning in psychiatry. Nat Neurosci. 2017
Oct;20(10):1342-1349. PMID: 28805813

TODO: Pending improvements:
- Add Panel C
- source the n from the column names for non stardard analyses
- use this script as guide: /lab/r.projects/COVID_drug_repurposing/4.drug_repurp_explanation_Fig.S5.R
}
\keyword{cdr}
\keyword{gtp}
\keyword{showcase}
\keyword{visualization}
