% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources_split_gctx.R
\name{split_gctx}
\alias{split_gctx}
\title{Split gctx files in RDS files}
\usage{
split_gctx(
  split.in.n = 300,
  parent.signature.dir =
    "/sc/arion/projects/va-biobank/resources/CMap/cmap_l1000_2021-11-20/",
  distil_cc_q75_min = 0.2,
  pct_self_rank_q25_max = 0.05,
  grep.sig.pattern = "trt_cp|trt_sh|trt_oe|trt_xpr"
)
}
\arguments{
\item{split.in.n}{How many signatures in each RDS file. Default is 300}

\item{parent.signature.dir}{Location of signature files. These are gw Consider saving locally to spead things up (e.g. /scratch/cmap_l1000_2021_01_28/)}

\item{distil_cc_q75_min}{75th quantile of pairwise Spearman correlations in landmark space of replicate level 4 profiles. Default is 0.2 corresponding to gold signatures.}

\item{pct_self_rank_q25_max}{Self connectivity of replicates expressed as a percentage of total instances in a replicate set. Default is 0.05 corresponding to gold signatures.}
}
\value{
N/A. Saves the figures
}
\description{
Prepare RDS files from LINCS gctx files for main pipeline. Need to run this for xpr
Requires 1) cmapR (either from bioconductor or github), 2) .gctx files, 3) siginfo_beta.txt
}
\details{
Defaults to selecting "Golden" signatures, a heuristic for assessing whether
a signature is reproducible and distinct:
https://docs.google.com/document/d/1q2gciWRhVCAAnlvF2iRLuJ7whrGP6QjpsCMq1yWz7dU
Requires cmapR package either from bioconductor or github
}
\keyword{miami}
\keyword{plot}
\keyword{regional}
\keyword{visualization}
